its fkin done, we need to refractor and improve performance

+++ src/CellsCalibrationFactors.h
@@ -9,29 +9,29 @@ class CellsCalibrationFactors
 public:
     CellsCalibrationFactors();
 
-    void set_first_cell(double first_cell);
-    double get_first_cell();
+    void set_first_cell(float first_cell);
+    float get_first_cell();
 
-    void set_second_cell(double second_cell);
-    double get_second_cell();
+    void set_second_cell(float second_cell);
+    float get_second_cell();
 
-    void set_third_cell(double third_cell);
-    double get_third_cell();
+    void set_third_cell(float third_cell);
+    float get_third_cell();
 
-    void set_fourth_cell(double fourth_cell);
-    double get_fourth_cell();
+    void set_fourth_cell(float fourth_cell);
+    float get_fourth_cell();
 
-    void set_first_read_1(double value);
-    double get_first_read_1();
+    void set_first_read_1(float value);
+    float get_first_read_1();
 
-    void set_first_read_2(double value);
-    double get_first_read_2();
+    void set_first_read_2(float value);
+    float get_first_read_2();
 
-    void set_first_read_3(double value);
-    double get_first_read_3();
+    void set_first_read_3(float value);
+    float get_first_read_3();
 
-    void set_first_read_4(double value);
-    double get_first_read_4();
+    void set_first_read_4(float value);
+    float get_first_read_4();
 
 private:
     Preferences preferences; // Instancia de la clase Preferences
